+++ pkg/client/config_test.go
@@ -45,6 +45,7 @@ telepresenceAPI:
   port: 1234
 intercept:
   appProtocolStrategy: portName
+  defaultPort: 9080
 `,
 	}
 
@@ -83,6 +84,7 @@ intercept:
 	assert.Equal(t, "ambassador-telepresence-webhook-image:0.0.2", cfg.Images.WebhookAgentImage) // from user
 	assert.Equal(t, 1234, cfg.TelepresenceAPI.Port)                                              // from user
 	assert.Equal(t, PortName, cfg.Intercept.AppProtocolStrategy)                                 // from user
+	assert.Equal(t, 9080, cfg.Intercept.DefaultPort)                                             // from user
 }
 
 func Test_ConfigMarshalYAML(t *testing.T) {
@@ -98,6 +100,7 @@ func Test_ConfigMarshalYAML(t *testing.T) {
 	cfg.Grpc.MaxReceiveSize, _ = resource.ParseQuantity("20Mi")
 	cfg.TelepresenceAPI.Port = 4567
 	cfg.Intercept.AppProtocolStrategy = PortName
+	cfg.Intercept.DefaultPort = 9080
 	cfgBytes, err := yaml.Marshal(cfg)
 	require.NoError(t, err)