+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### 2.4.10 (TBD)
 
+- Feature: The port used by default in the `telepresence intercept` command (8080), can now be changed by setting
+  the `intercept.defaultPort` in the `config.yml` file.
+
 - Feature: The strategy when selecting the application protocol for personal intercepts in agents injected by the 
   mutating webhook can now be configured using the `agentInjector.appProtocolStrategy` in the Helm chart.